+++ Basil/include/Scene/Component.h
@@ -8,6 +8,8 @@
 
 #include <glm/glm.hpp>
 #include <glm/gtc/matrix_transform.hpp>
+#define GLM_ENABLE_EXPERIMENTAL
+#include <glm/gtx/quaternion.hpp>
 
 namespace Basil
 {
@@ -59,9 +61,7 @@ namespace Basil
 		
 		glm::mat4 getTransform() const
 		{
-			glm::mat4 localRotation = glm::rotate(glm::mat4(1.0f), rotation.x, { 1, 0, 0 })
-									* glm::rotate(glm::mat4(1.0f), rotation.y, { 0, 1, 0 })
-									* glm::rotate(glm::mat4(1.0f), rotation.z, { 0, 0, 1 });
+			glm::mat4 localRotation = glm::toMat4(glm::quat(rotation));
 
 			return glm::translate(glm::mat4(1.0f), translation)
 					* localRotation